@@ -13,30 +13,31 @@ import (
 	"encoding/json"
 	_ "expvar"
 	"flag"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
 	"path/filepath"
 	"runtime"
 	"runtime/pprof"
+	"strings"
 	"time"
 
 	"github.com/blevesearch/bleve/v2"
 	bleveHttp "github.com/blevesearch/bleve/v2/http"
 )
 
-var batchSize = flag.Int("batchSize", 100, "batch size for indexing")
-var bindAddr = flag.String("addr", ":8094", "http listen address")
-var jsonDir = flag.String("jsonDir", "data/", "json directory")
-var indexPath = flag.String("index", "beer-search.bleve", "index path")
-var staticEtag = flag.String("staticEtag", "", "A static etag value.")
-var staticPath = flag.String("static", "static/", "Path to the static content")
-var cpuprofile = flag.String("cpuprofile", "", "write cpu profile to file")
-var memprofile = flag.String("memprofile", "", "write mem profile to file")
+var (
+	batchSize  = flag.Int("batchSize", 100, "batch size for indexing")
+	bindAddr   = flag.String("addr", ":8094", "http listen address")
+	jsonDir    = flag.String("jsonDir", "data/", "json directory")
+	indexPath  = flag.String("index", "beer-search.bleve", "index path")
+	staticEtag = flag.String("staticEtag", "", "A static etag value.")
+	staticPath = flag.String("static", "static/", "Path to the static content")
+	cpuprofile = flag.String("cpuprofile", "", "write cpu profile to file")
+	memprofile = flag.String("memprofile", "", "write mem profile to file")
+)
 
 func main() {
-
 	flag.Parse()
 
 	log.Printf("GOMAXPROCS: %d", runtime.GOMAXPROCS(-1))
@@ -46,7 +47,10 @@ func main() {
 		if err != nil {
 			log.Fatal(err)
 		}
-		pprof.StartCPUProfile(f)
+
+		if err := pprof.StartCPUProfile(f); err != nil {
+			log.Fatal(err)
+		}
 	}
 
 	// open the index
@@ -75,7 +79,11 @@ func main() {
 				if err != nil {
 					log.Fatal(err)
 				}
-				pprof.WriteHeapProfile(f)
+
+				if err := pprof.WriteHeapProfile(f); err != nil {
+					log.Fatal(err)
+				}
+
 				f.Close()
 			}
 		}()
@@ -101,15 +109,23 @@ func main() {
 
 	// start the HTTP server
 	http.Handle("/", router)
-	log.Printf("Listening on %v", *bindAddr)
-	log.Fatal(http.ListenAndServe(*bindAddr, nil))
 
+	address := strings.Split(*bindAddr, ":")
+	host := address[0]
+	port := address[1]
+
+	if host == "" {
+		host = "localhost"
+	}
+
+	log.Printf("Listening on http://%v:%v", host, port)
+
+	log.Fatal(http.ListenAndServe(*bindAddr, nil))
 }
 
 func indexBeer(i bleve.Index) error {
-
 	// open the directory
-	dirEntries, err := ioutil.ReadDir(*jsonDir)
+	dirEntries, err := os.ReadDir(*jsonDir)
 	if err != nil {
 		return err
 	}
@@ -123,7 +139,7 @@ func indexBeer(i bleve.Index) error {
 	for _, dirEntry := range dirEntries {
 		filename := dirEntry.Name()
 		// read the bytes
-		jsonBytes, err := ioutil.ReadFile(*jsonDir + "/" + filename)
+		jsonBytes, err := os.ReadFile(*jsonDir + "/" + filename)
 		if err != nil {
 			return err
 		}
@@ -135,7 +151,10 @@ func indexBeer(i bleve.Index) error {
 		}
 		ext := filepath.Ext(filename)
 		docID := filename[:(len(filename) - len(ext))]
-		batch.Index(docID, jsonDoc)
+		if err := batch.Index(docID, jsonDoc); err != nil {
+			return err
+		}
+
 		batchCount++
 
 		if batchCount >= *batchSize {
